requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0.post2
Jinja2==3.1.2
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.1.2
